pyproject.toml
[project]
name = "kakuro"
version = "0.1.0"
description = "High-performance Kakuro puzzle generator with C++ acceleration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.126.0",
    "numpy>=2.4.0",
    "pydantic[email]>=2.12.5",
    "pyinstaller>=6.17.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.1.0",
    "starlette[full]>=0.50.0",
    "PyJWT>=2.8.0",
    "email-validator>=2.1.0",
    "requests>=2.32.5",
    "pybind11>=3.0.1",
    "pytest>=9.0.2",
]

[build-system]
requires = [
    "setuptools>=45",
    "wheel",
    "pybind11>=2.11.0",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["python", "routes"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.so", "*.pyd", "*.dll"]
